stm32h743xx_udb: Replace UDB flag with generic flag
TODO(https://github.com/ARMmbed/DAPLink/issues/963):
MSC sometimes fails for unknown reason on stm32h743. We disabled mass
storage as a workaround for it earlier. Repalce the previous UDB flag
with sth more generic. This is only an issue for UDB because UDB is
configured to crash on asserts.